% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vectors.R
\name{check_binning}
\alias{check_binning}
\title{helper function to check the input of binning}
\usage{
check_binning(bin_param, bin_type, w_x, w_y)
}
\arguments{
\item{bin_param}{A numeric vector indicating the size of the bin. If the
\code{bin_type} is "square" or "rectangle", this will be a vector of length
two giving the numbers of rectangular quadrats in the x and y directions. If
the \code{bin_type} is "hexagonal", this will be a number giving the side
length of hexagons. Positive numbers only.}

\item{bin_type}{A string indicating which bin shape is to be used for
vectorization. One of "square" (default), "rectangle", or "hexagon".}

\item{w_x}{A numeric vector of length two specifying the x coordinate
limits of enclosing box.}

\item{w_y}{A numeric vector of length two specifying the y coordinate
limits of enclosing box.}
}
\value{
the length of total bins
}
\description{
helper function to check the input of binning
}
