% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genesets.R
\name{check_geneset_input}
\alias{check_geneset_input}
\title{helper function to check the inputs passed to create geneset function}
\usage{
check_geneset_input(data_lst, bin_type, bin_param, w_x, w_y, cluster_info)
}
\arguments{
\item{data_lst}{A list of named matrices containing the coordiantes of 
transcripts. This can be the output from the function \code{\link{get_data}}.}

\item{bin_type}{A string indicating which bin shape is to be used for
vectorization. One of "square" (default), "rectangle", or "hexagon".}

\item{bin_param}{A numeric vector indicating the size of the bin. If the
\code{bin_type} is "square" or "rectangle", this will be a vector of length
two giving the numbers of rectangular quadrats in the x and y directions. If
the \code{bin_type} is "hexagonal", this will be a number giving the side
length of hexagons. Positive numbers only.}

\item{w_x}{A numeric vector of length two specifying the x coordinate
limits of enclosing box.}

\item{w_y}{A numeric vector of length two specifying the y coordinate
limits of enclosing box.}

\item{cluster_info}{A dataframe/matrix containing the centroid coordinates,
cluster sample label for each cell.The column names must include
"x" (x coordinate), "y" (y coordinate),
"cluster" (cluster label) and "sample" (sample).}
}
\value{
A list of two elements
}
\description{
helper function to check the inputs passed to create geneset function
}
