% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceMarker-package.R
\docType{data}
\name{gene_info}
\alias{gene_info}
\title{Gene panel information}
\format{
A data frame with 313 rows and 3 variables:
\describe{
    \item{Name}{gene name}
    \item{Ensembl.ID}{id}
    \item{Annotation}{annotation label of the gene}
}
}
\source{
\url{https://cf.10xgenomics.com/samples/xenium/1.0.1/
Xenium_FFPE_Human_Breast_Cancer_Rep1/
Xenium_FFPE_Human_Breast_Cancer_Rep1_panel.tsv}
}
\usage{
data(gene_info)
}
\value{
data frame
}
\description{
A csv file containing the information for the gene panel
}
\keyword{datasets}
