% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vectors.R
\name{get_cluster_vectors}
\alias{get_cluster_vectors}
\title{Create spatial vectors for clusters}
\usage{
get_cluster_vectors(cluster_info, bin_length, bin_type, bin_param, w_x, w_y)
}
\arguments{
\item{cluster_info}{A dataframe/matrix containing the centroid coordinates,
cluster label and sample for each cell.The column names must include
"x" (x coordinate), "y" (y coordinate),
"cluster" (cluster label) and "sample" (sample).}

\item{bin_length}{A positive integer giving the length of total bins}

\item{bin_type}{A string indicating which bin shape is to be used for
vectorization. One of "square" (default), "rectangle", or "hexagon".}

\item{bin_param}{A numeric vector indicating the size of the bin. If the
\code{bin_type} is "square" or "rectangle", this will be a vector of length
two giving the numbers of rectangular quadrats in the x and y directions. If
the \code{bin_type} is "hexagonal", this will be a number giving the side
length of hexagons. Positive numbers only.}

\item{w_x}{A numeric vector of length two specifying the x coordinate
limits of enclosing box.}

\item{w_y}{A numeric vector of length two specifying the y coordinate
limits of enclosing box.}
}
\value{
a matrix contains the cell count in each grid.
Each row refers to a grid, and each column refers to a cluster.
}
\description{
Create spatial vectors for clusters
}
