% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read 10x Xenium data.}
\usage{
get_data(
    path,
    mtx_name,
    trans_name = "transcript_info.csv.gz",
    cells_name = "cell_info.csv.gz"
)
}
\arguments{
\item{path}{The directory containing the matrix.mtx,
genes.tsv (or features.tsv), and barcodes.tsv files provided by 10X.}

\item{mtx_name}{The folder name under path containing the matrix.mtx,
genes.tsv (or features.tsv), and barcodes.tsv files.}

\item{trans_name}{The file name under path containing the transcript
information files. Default is "transcript_info.csv.gz"}

\item{cells_name}{The file name under path containing the cell information
files. Default is "cell_info.csv.gz"}
}
\value{
A named list with the following components
\item{\code{cm}  }{ The count matrix. Each row refers to a gene,
and each column refers to a cell}
\item{\code{cm_neg}  }{ The count matrix for negative control
probe and codeword genes. Each row refers to a probe/codeword gene,
and each column refers to a cell.}
\item{\code{trans_info}  }{A matrix contains the transcript information
for each transcript. All the columns read from the input \code{trans_name}
path is kept.}
\item{\code{cell_info}  }{A matrix contains the cell information
for each cell. The coordinates of each cell is multiplied with the input
\code{pixel_size}. All the columns read from the input \code{cells_name}
path is kept.}
\item{\code{zero_cells}  }{A vector giving the names of cell with zero count
on every cell. }
\item{\code{probe}  }{A vector giving the names of negative control probe
genes.}
\item{\code{codeword}  }{A vector giving the names of negative control
codeword genes.}
}
\description{
This function will read the 10x Xenium data and return a list of matrices
and vectors. The coordinates will be in microns.
}
\examples{
\dontrun{
path <- 'path/to/data/directory'
data = get_data(path, mtx_name="cell_feature_matrix",
                trans_name="transcripts.csv.gz",
                cells_name="cells.csv.gz")
                }
}
