% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vectors.R
\name{get_gene_vectors_cm}
\alias{get_gene_vectors_cm}
\title{Create spatial vectors for genes from count matrix and cell coordinates}
\usage{
get_gene_vectors_cm(
    cluster_info,
    cm_lst,
    bin_type,
    bin_param,
    all_genes,
    w_x,
    w_y
)
}
\arguments{
\item{cluster_info}{A dataframe/matrix containing the centroid coordinates,
cluster label and replicate for each cell.The column names must include
"x_centroid" (x coordinate), "y_centroid" (y coordinate),
"cluster" (cluster label) and "rep" (replicate).}

\item{cm_lst}{A list of named matrices containing the count matrix for
each replicate. The name must match the rep column in \code{cluster_info}.
If this input is provided, the \code{cluster_info} must be specified and
contain an additional column "cell_id" to link cell location and
count matrix. Default is NULL.}

\item{bin_type}{A string indicating which bin shape is to be used for
vectorization. One of "square" (default), "rectangle", or "hexagon".}

\item{bin_param}{A numeric vector indicating the size of the bin. If the
\code{bin_type} is "square" or "rectangle", this will be a vector of length
two giving the numbers of rectangular quadrats in the x and y directions. If
the \code{bin_type} is "hexagonal", this will be a number giving the side
length of hexagons. Positive numbers only.}

\item{all_genes}{A vector of strings giving the name of the genes you want to
test. This will be used as column names for one of the result matrix
\code{gene_mt}.}

\item{w_x}{A numeric vector of length two specifying the x coordinate
limits of enclosing box.}

\item{w_y}{A numeric vector of length two specifying the y coordinate
limits of enclosing box.}
}
\value{
a matrix contains the transcript count in each grid.
Each row refers to a grid, and each column refers to a gene.
}
\description{
This function will build gene vectors with count matrix and cell locations
}
