% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koalasso-package.R
\docType{data}
\name{rep2_clusters}
\alias{rep2_clusters}
\title{Rep2 selected cells}
\format{
A data frame with 1815 rows and and 6 variables:
\describe{
    \item{anno}{the provided cell type annotation}
    \item{cluster}{cluster label}
    \item{x}{x coordinates}
    \item{y}{y coordiantes}
    \item{cells}{cell id}
    \item{sample}{sample id}
}
}
\source{
\url{https://cf.10xgenomics.com/samples/xenium/1.0.1/
Xenium_FFPE_Human_Breast_Cancer_Rep2/
Xenium_FFPE_Human_Breast_Cancer_Rep2_outs.zip}
}
\usage{
data(rep2_clusters)
}
\value{
data frame
}
\description{
A csv file containing the coordinates and cluster label for each cell of the
selected subset of rep2.
}
\keyword{datasets}
