% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceMarker-package.R
\docType{data}
\name{rep2_neg}
\alias{rep2_neg}
\title{Rep2 negative control genes within the selected region.}
\format{
A list of named matrices:
\describe{
    \item{cm}{the count matrix for negative control genes}
    \item{trans_info}{a data frame containing the coordinates for 
                        every transcript}
    \item{probe}{names of negative control probe genes}
    \item{codeword}{names of negative control codeword genes}
}
}
\source{
\url{https://cf.10xgenomics.com/samples/xenium/1.0.1/
Xenium_FFPE_Human_Breast_Cancer_Rep2/
Xenium_FFPE_Human_Breast_Cancer_Rep2_outs.zip}
}
\usage{
data(rep2_neg)
}
\value{
List
}
\description{
A list of named matrices
}
\keyword{datasets}
