% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koalasso-package.R
\docType{data}
\name{rep2_sub}
\alias{rep2_sub}
\title{A small section of Xenium human breast cancer rep2.}
\format{
A list of named matrices generated from get_data():
\describe{
    \item{cm}{the count matrix}
    \item{trans_info}{a data frame containing the coordiantes 
                     for every transcript}
}
}
\source{
\url{https://cf.10xgenomics.com/samples/xenium/1.0.1/
Xenium_FFPE_Human_Breast_Cancer_Rep2/
Xenium_FFPE_Human_Breast_Cancer_Rep2_outs.zip}
}
\usage{
data(rep2_sub)
}
\value{
List
}
\description{
A list of named matrices generated from get_data()
}
\keyword{datasets}
